Get the package build working
This requires figuring out a valid version for the runtime.  We were
started with master, but that's apparently invalid.  (Okay....) I
next guessed 42, since I'm on Gnome 42, but this failed for unclear
reasons.  But 45 is the most recent (?) and appears to work.  If
this is documented anywhere, I failed to find it.

Although the program worked when run from Builder, it failed with
installed with Flatpak.  Eventually, I traced the problem to the
pygments Python package, which is required by nbconvert.  It was
around during the build step, so pip didn't install it, but it
doesn't get included in the runtime container.  (WTF?)  So we
just copy it over manually.

At this point you can build the flatpak package, install it with
`flatpak install`, and run it with `flatpack run`.