Register as an opener for notebook files
In the desktop file, add a %F indicating file names can be passed in.
Also, set the mimetype to indicate which types of files to recommend
this for.  (This bit doesn't seem to be working?)

In the application, add a do_open handler.  Documentation [1] suggests
that it gets two arguments, a list of file and a hint (a string, likely
empty).  But it gets a third argument, between those two, probably the
number of files being passed in.  It doesn't matter; we can ignore both
of them.

[1] https://lazka.github.io/pgi-docs/Gio-2.0/classes/Application.html#Gio.Application.do_open

This allows files to be opened from Nautilus as well as from the command
line, with the `open` or `xdg-open` commands.