Show images with relative URIs
To make this work, we can set the base URI in the load_html function.
In the default Flatpak configuration, filesystem access is mediated
and the file path we actually get is a sandbox /var/run/....  We
can get the actual filename with the "-filesystem=host" option in
the build args.  (Note that we cannot add the ":ro" suffix to this
for read-only access, since that again gives us sandboxed paths.)

Note that this doesn't work when run from within Gnome Builder -- we
still get sandboxed paths here.  It only works when the Flatpak bundle
is installed.  Presumably there's a different set of build arguments
that Gnome Builder looks at, but they do such a good job of hiding the
configuration details that I haven't found it.